+++ src/main/java/com/pfe/elearning/repository/CandidateRepository.java
@@ -4,5 +4,5 @@ import com.pfe.elearning.entity.Candidate;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CandidateRepository extends JpaRepository<Candidate, Integer> {
+public interface CandidateRepository extends JpaRepository<Candidate, Long> {
 }
